+++ docs/api/source-files.js
@@ -1,5 +1,0 @@
-var N = null;var sourcesIndex = {};
-sourcesIndex["cargo_make"] = {"name":"","files":["main.rs"]};
-sourcesIndex["cli"] = {"name":"","dirs":[{"name":"cli_commands","files":["diff_steps.rs","list_steps.rs","mod.rs","print_steps.rs"]},{"name":"environment","files":["crateinfo.rs","mod.rs"]},{"name":"functions","files":["decode_func.rs","mod.rs","remove_empty_func.rs","split_func.rs","trim_func.rs"]},{"name":"installer","files":["cargo_plugin_installer.rs","crate_installer.rs","crate_version_check.rs","mod.rs","rustup_component_installer.rs"]},{"name":"scriptengine","files":["duck_script.rs","generic_script.rs","mod.rs","os_script.rs","rsscript.rs","script_utils.rs","shebang_script.rs","shell_to_batch.rs"]}],"files":["cache.rs","cli.rs","command.rs","condition.rs","config.rs","descriptor.rs","execution_plan.rs","io.rs","legacy.rs","logger.rs","mod.rs","profile.rs","runner.rs","storage.rs","toolchain.rs","types.rs","version.rs"]};
-sourcesIndex["makers"] = {"name":"","files":["makers.rs"]};
-createSourceSidebar();