+++ src/lib/scriptengine/duck_script.rs
@@ -7,9 +7,11 @@
 #[path = "./duck_script_test.rs"]
 mod duck_script_test;
 
+use crate::environment;
 use duckscript::runner;
 use duckscript::types::runtime::Context;
 use duckscriptsdk;
+use envmnt;
 
 pub(crate) fn execute(script: &Vec<String>, cli_arguments: &Vec<String>, validate: bool) {
     let script_text = script.join("\n");
@@ -25,17 +27,28 @@ pub(crate) fn execute(script: &Vec<String>, cli_arguments: &Vec<String>, validat
     }
 
     match duckscriptsdk::load(&mut context.commands) {
-        Ok(_) => match runner::run_script(&script_text, context) {
-            Ok(_) => (),
-            Err(error) => {
-                if validate {
-                    error!("Error while running duckscript: {}", error)
+        Ok(_) => {
+            let directory = envmnt::get_or("CARGO_MAKE_WORKING_DIRECTORY", "");
+
+            match runner::run_script(&script_text, context) {
+                Ok(_) => (),
+                Err(error) => {
+                    if validate {
+                        error!("Error while running duckscript: {}", error);
+                        panic!("Error while running duckscript: {}", error)
+                    }
                 }
+            };
+
+            // revert to originl working directory
+            if !directory.is_empty() {
+                environment::setup_cwd(Some(&directory));
             }
-        },
+        }
         Err(error) => {
             if validate {
-                error!("Unable to load duckscript SDK: {}", error)
+                error!("Unable to load duckscript SDK: {}", error);
+                panic!("Unable to load duckscript SDK: {}", error)
             }
         }
     };